提交后台修改代码

+++ src/main/java/com/oscar/migration/service/impl/SysRoleServiceImpl.java
@@ -106,10 +106,9 @@ public class SysRoleServiceImpl implements SysRoleService {
         Optional<SysRole> op = sysRoleRepository.findById(sysRole.getId());
         if (op.isPresent()) {
             SysRole echoRole = op.get();
-//            TODO
-//            if (SysConstants.ADMIN_NAME.equals(echoRole.getName())) {
-//                return Result.error("管理员角色，不允许修改！");
-//            }
+            if (SysConstants.ADMIN_NAME.equals(echoRole.getName())) {
+                return Result.error("管理员角色，不允许修改！");
+            }
             sysRole.setLaseUpdateTime(new Date());
             sysRoleRepository.save(sysRole);
             /**更新角色菜单关系*/
@@ -168,10 +167,9 @@ public class SysRoleServiceImpl implements SysRoleService {
         }
         Optional<SysRole> role = sysRoleRepository.findById(roleId);
         if (role.isPresent()) {
-            //TODO
-//            if (SysConstants.ADMIN_NAME.equals(role.get().getName())) {
-//                return Result.error("管理员拥有所有菜单权限，不允许修改！");
-//            }
+            if (SysConstants.ADMIN_NAME.equals(role.get().getName())) {
+                return Result.error("管理员拥有所有菜单权限，不允许修改！");
+            }
             deleteRoleMenus(roleId);
             List<SysRoleMenu> list = new ArrayList<>();
             Date date = new Date();
